requirements.txt
boto==2.44.0
boto3==1.4.2
Django==1.10.5
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
django-cors-headers==1.3.1
django-filter==1.0.1
django-storages==1.5.1
gunicorn==19.6.0
PyMySQL==0.7.9
PyJWT==1.4.2
